requirements.txt
# aiohappyeyeballs==2.4.0
# aiohttp==3.10.5
# aiosignal==1.3.1
# annotated-types==0.7.0
# anyio==4.4.0
# attrs==24.2.0
# beautifulsoup4==4.12.3
# bs4==0.0.2
# certifi==2024.8.30
# charset-normalizer==3.3.2
# dataclasses-json==0.6.7
# distro==1.9.0
# faiss-cpu==1.8.0.post1
# frozenlist==1.4.1
# h11==0.14.0
# httpcore==1.0.5
# httpx==0.27.2
# idna==3.8
# jiter==0.5.0
# jsonpatch==1.33
# jsonpointer==3.0.0
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.37
langchain-experimental==0.0.64
langchain-openai==0.1.23
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langchainhub==0.1.21
langsmith==0.1.108
# marshmallow==3.22.0
# multidict==6.0.5
# mypy-extensions==1.0.0
# numpy==1.26.4
openai==1.43.0
# orjson==3.10.7
# packaging==24.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
# pydantic==2.8.2
# pydantic_core==2.20.1
# pypdf==4.3.1
python-dotenv==1.0.1
# PyYAML==6.0.2
# regex==2024.7.24
# requests==2.32.3
# sniffio==1.3.1
# soupsieve==2.6
# SQLAlchemy==2.0.32
supabase==2.11.0
# tenacity==8.5.0
# tiktoken==0.7.0
# tqdm==4.66.5
# types-requests==2.32.0.20240712
# typing-inspect==0.9.0
# typing_extensions==4.12.2
# urllib3==2.2.2
# yarl==1.9.7
